user-service/requirements.txt
fastapi
uvicorn
pymongo
python-jose
python-dotenv
email-validator
